Remove references to 'instance_type'
They haven't been instance types in a very long time. We're stuck with
that terminology in the database for now, but otherwise we can replace
references to 'instance_type' with 'flavor'.

While we're here, we also rename a single case of an 'i' variable to a
more informative 'idx'. 'i' is too short to be useful outside of a very
limited scope.

Change-Id: I81fec10535034f3a81d46713a6eda813f90561cf
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>