api: Log os-resetState as an instance action
It appears this was missed as this action never hits the compute API and
is fully handled within the API itself. This change simply records an
instance action directly in the API for this admin action providing a
breadcrumb to operators and users.

Closes-Bug: #1911924
Change-Id: Ifd9657f3357bc39cb904caa65168d38def8c9184
(cherry picked from commit 534b4f38958af2fbf4392e445ddb1b2bad4179ed)