docs: Move the LibvirtDistroSupportMatrix wiki page into our docs
This change moves the LibvirtDistroSupportMatrix [1] wiki page into the
tree as a reference doc. The wiki page will be decommissioned once this
change lands and is published.

Some older distro information is removed to keep the table readable and
a note is added to driver.py to ensure it updated with each version
bump.

[1] https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

Change-Id: Id49a4e400159130fbc676800aeca6b9746071a2e