libvirt: add AsyncDeviceEventsHandler
This class will be used as synchronization point between the device
detach implementation and the generic libvirt event handling in the
libvirt driver. This later makes it possible to wait for libvirt device
removal events during device detach in a thread / greenlet safe way.

The basic usage pattern is the following:

setup:
  handler = AsyncDeviceEventsHandler()

thread1:

  waiter = handler.create_waiter(<instance>, <device>, <event types>)
  # initiate detach in libvirt
  detach()
  # block until the event is received
  event = handler.wait(waiter, timeout=20)

thread2:

  # at some point after detach() in thread1
  handler.notify_waiters(event)

This is part of the longer series trying to transform the existing
device detach handling to use libvirt events.

Change-Id: I7fc1ba2d2cbdb3f46f3649b1279b0c1a40457647
Related-Bug: #1882521