libvirt: parse alias out from device config
Libvirt assigns a device alias for each device in the domain and uses
this identifier in the events related to these devices, like device
detach.

This patch extends the LibvirtConfigGuestInterface and the
LibvirtConfigGuestDisk config object to parse the alias back from the
domain config. The alias is intentionally not used during domain XML
generation as it is expected that libvirt will populate and persist
this field.

This is part of the longer series trying to transform the existing
device detach handling to use libvirt events.

Change-Id: I16e7df6932bb7dff243706ee49338ba6b3782085
Related-Bug: #1882521