zuul: Mark nova-lvm as voting
As covered in the removed TODO this was to be done after bug #1771700
was resolved in Queens but instability in the job over the years stopped
this from ever happening.

With I7ad1d9820c48644bff7834560cdc7d5c13f7ed41 hopefully addressing the
only outstanding stability issue we can now finally mark the job as
voting again.

Related-Bug: #1771700
Change-Id: Ie9cba36ab602ab7c20c9a2c3b60e11e8506581ce