Remove unused 'shutdown_terminate' flag
This seems to be a hangover from the days of EC2. There are no callers
relying setting this flag except for tests, meaning we can remove it. We
keep the code that checks for this value since it's conceivable that
there are very old instances out there with it set, but this too can be
removed in the distant future.

Change-Id: I782059aa2933aa414b94724f1e61f48466bb7a04
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>