Restart glance after updating policy in job
When we write out the glance policy for the multistore job, we do
so after glance has already started. After the json->yaml change,
we will decide too early in g-api lifetime that the legacy file
isn't present and thus never honor it later. So, create it and then
restart glance services.

Change-Id: Ic1c01366dbfdcfb85750b85f960b76aea934db59