Centralize sqlite FK constraint enforcement
There's a TODO in the code to centralize foreign key constraint
enforcement for sqlite for unit and functional tests and we're missing
enforcement of FK constraints in a couple of test classes that should
have it.

This resolves the TODO and turns on FK constraint enforcement where it
is missing. Do this to enhance testing in preparation for a proposed
change to the database archiving logic later in this patch series.

Conflicts:
    nova/test.py

NOTE(melwitt): The conflicts are because of the following changes not
in Victoria:

  * Ib2c406327fef2fb4868d8050fc476a7d17706e23 (Remove six.moves)
  * Ide65686cf02463045f5c32771ca949802b19636f (Remove
    six.binary_type/integer_types/string_types)

Change-Id: Idcf026d020e63e4e6ece1db46e4cdc7b7742b76f
(cherry picked from commit 172024db7174bdac05a3d3172c645b0db83ae41e)