Centralize sqlite FK constraint enforcement
There's a TODO in the code to centralize foreign key constraint
enforcement for sqlite for unit and functional tests and we're missing
enforcement of FK constraints in a couple of test classes that should
have it.

This resolves the TODO and turns on FK constraint enforcement where it
is missing. Do this to enhance testing in preparation for a proposed
change to the database archiving logic later in this patch series.

Change-Id: Idcf026d020e63e4e6ece1db46e4cdc7b7742b76f
(cherry picked from commit 172024db7174bdac05a3d3172c645b0db83ae41e)
(cherry picked from commit 382d64ea36e5cf3b47929fee0b93539d2f126975)