fup: Merge duplicate volume attachment checks
This is a follow up to I2881d77d52bcbde9f3ac6a6ddfb4a22a9bd45c8a where
it was noted that we could merge these checks into one using a single db
query for all active volume attachments for a given volume. This
requires changes to the BlockDeviceMappingList object and was thus left
as a separate change to allow the original to be backported.

Change-Id: I859be652eb8fff591899a223193bd9efc8830cf3