[ironic] correct capabilities for disk images supported
Generally ironic users use whole disk images or partition images,
and ironic identifies the difference between the two by checking
if a kernel and ramdisk are in use for the system.

Turns out that an AMI is a partition image, where a kernel and
ramdisk is paired with the desired disk image. So in essence
we appear to have supported it all along. Just different names were
used.

It *appears* that this should work just fine in most cases unless
someone has somewhat weird settings on their ironic deployment,
and realistically those should be considered bugs if they are
encountered.

Change-Id: I168d75da5c04270459ce80d913206910f79bacb6