libvirt: Always enable USB controller on PPC64
In change I5b16d9ddc9cad56b367d1ad94abba95d675840b0, we started
disabling the USB controller whenever we deemed it unnecessary. We did
this by scanning through the requested devices and identifying those
that used the USB bus, only adding the controller if one or more
matching devices were found.

It transpires that just as QEMU adds PS2 input devices for x86, libvirt
automatically adds USB input devices for PPC64 [1]. These devices won't
appear in the XML we provide to libvirt, since we, the user, are not the
ones requesting them, but the disabling of the instance will result in
the instance failing to boot and libvirt emitting the following error
message:

  unsupported configuration: USB is disabled for this domain, but USB
  devices are present in the domain XML

Resolve this issue by bypassing the devices check and leaving
configuration of the USB controller to libvirt, as before.

[1] https://github.com/libvirt/libvirt/blob/3d42a5766/src/qemu/qemu_domain.c#L3559-L3560

Change-Id: Iefcb4bbe1ae31dcf79a3ba88372b4971d97ad2ea
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes-Bug: #1914259