Add net & utils methods for routed nets & segments
Before adding a prefilter, we need to add new methods for asking Neutron
about the related aggregates we need to ask Placement if the user
creates or moves an instance with a requested network or a port.

NOTE(sbauza): Looks like mypy doesn't like the current methods in
nova.network.neutron, so we need to fix the issues before adding this
module to mypy-files.txt.

Partially-Implements : blueprint routed-networks-scheduling

Change-Id: Ie166f3b51fddeaf916cda7c5ac34bbcdda0fd17a