Differentiate between InstanceNotFound and ConstraintNotMet
In the DB API instance_destroy method, we have logic that assumes that
if a soft_delete() call returns 0 rows, it must mean that a query
constraint was not met and it raises ConstraintNotMet.

There is however another reason soft_delete() can return 0 rows: if the
instance record has already been deleted by a racing request. In this
case, raising ConstraintNotMet is wrong and could cause incorrect
exception handling or be misleading when trying to debug issues.

This adds a check for the existence of the instance record when
soft_delete() returns 0 so that InstanceNotFound will be raised if the
instance record is gone, else ConstraintNotMet will be raised.

Related-Bug: #1914777
Closes-Bug: #1917945

Change-Id: Ic2d5a1ec53d2f787907deafbf598ba581e6f2f65