nova-next: Start testing the q35 machine type
With the introduction of the libvirt-default-machine-type blueprint
operators are now able to change the default machine type used by an
existing environment more easily. This should hopefully allow more
deployments to switch to the modern q35 machine type, away from the
legacy pc machine type that is currently the default.

With this in mind this change aims to start testing the q35 machine type
as part of the nova-next job before we eventually change the default in
the future.

The following changes are made to the job to allow it to pass with q35:

- [libvirt]num_pcie_ports is increased to 24

This is required as more devices are now connected by default over PCIe.

- [compute-feature-enabled]ide_bus is disabled

The q35 machine type does not support an IDE bus so any tests using this
bus are now skipped.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/779179
Depends-On: https://review.opendev.org/c/openstack/tempest/+/775630
blueprint: libvirt-default-machine-type
Change-Id: Iad1adbc23b31dd54a96299e7a8a4b622c15eed8d