libvirt: Add firmware auto-configuration support
This is the guest or "domain" side of the config changes.

As noted in change Ib331f0ba3615dfe25f51cd9c948c97ec21c146d6, since
libvirt 5.2 [1], libvirt has provided the ability to automatically
configure bootloader paths. This means libvirt can now take a minimal
loader configuration like so:

  <os firmware='efi'>
    <type machine='q35'>hvm</type>
    <loader secure='yes'/>
    <boot dev='hd'/>
  </os>

and convert it into something explicit like so:

  <os>
    <type arch='x86_64' machine='pc-q35-5.1'>hvm</type>
    <loader readonly='yes' secure='yes' type='pflash'>{...}/OVMF_CODE.secboot.fd</loader>
    <nvram template='{...}/OVMF_VARS.secboot.fd'>{...}/mydomain_VARS.fd</nvram>
    <boot dev='hd'/>
  </os>

In change Ib331f0ba3615dfe25f51cd9c948c97ec21c146d6, we added the
ability to identify this capability from the host. In this change, we
add the ability to configure it for the guest.

[1] https://www.libvirt.org/news.html#v5-2-0-2019-04-03

Blueprint: allow-secure-boot-for-qemu-kvm-guests
Change-Id: I342e95fef0a792299180a4d8076e6e7c4041ac9b
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>