db: Fold in indexes
This will also make the alembic move simpler. Most of this change was
accomplished with "s/[a-z_]\+\.c\.\([a-z_]\+\)/'\1'/" (i.e. replace
"model.c.column" with "'column'"). The only other odd change is the use
of the 'mysql_length' parameter [1] to allow us to specify the length of
index entries for the 'migrations_by_host_nodes_and_status_idx' index on
the 'migrations' table.

[1] https://docs.sqlalchemy.org/en/14/dialects/mysql.html#index-length

Change-Id: Icad92d1b3e3a2e2cdd9cfb0d3f80ee091c697756
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>