db: Fold in ForeignKey constraints
Yet another thing to ease the path of the alembic migration. Once again,
there are some odd things here that require explanation. Firstly, we
need to work around the fact that some foreign keys are named when using
a MySQL or SQLite backend but not when using PostgreSQL. This is
addressed via a custom compile rule [1] that simply drops the name
argument for PostgreSQL, based on the assumption that we will not start
naming these constraints going forward (this seems sane given we've
stopped doing it for MySQL and SQLite). I had initially attempted to
solve this using dialect-specific keyword arguments, specifically, by
using the 'argument_for' function [2] to add a 'mysql_name' parameter,
but that doesn't seem to work as expected [3]. In addition to this, the
two foreign keys that were added without specific names ended up using
the sqlalchemy-migrate derived names of '%(table)s_%(firstcolumn)s_fkey'
[4] which we must replicate to prevent schema changes.

[1] https://docs.sqlalchemy.org/en/14/core/compiler.html
[2] https://docs.sqlalchemy.org/en/14/core/constraints.html#sqlalchemy.schema.ColumnCollectionConstraint.argument_for
[3] https://github.com/sqlalchemy/sqlalchemy/issues/5944
[4] https://opendev.org/x/sqlalchemy-migrate/src/commit/5d1f322542cd8eb42381612765be4ed9ca8105ec/migrate/changeset/constraint.py#L126-L139

Change-Id: I794cb5ed277b33e90087221b38b287e1132b639a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>