vmware: Handle folder renames in rescue cmd
When a VM is storage-vMotioned, the name of the folder its files are in
can change. Then, a rescue command trying to put the rescue disk into a
folder named after the instance's UUID cannot work anymore and actually
raises a FileNotFoundException for the directory path.

To fix this, we now take the root VMDK's folder and copy the rescue
image into that.

Change-Id: Icef785b96e51942e7bac2df10c116078c77fedc4
Closes-Bug: #1915815