Fixes the elapsed time logged during a live migration
Nova doesn’t log the actual time that it is taking to live
migrate an instance but the number of cycles/2 that it does
when checking the migration job. If any operation takes
longer inside the loop, the reported time is wrong.

This behavior can cause some confusion when operators are
debugging issues.

This patch ensures that Nova logs the elapsed time of the
live migration.

Closes-Bug: #1916031
Change-Id: I1d622c28a09ddd2aa7d33fa7057b2f78dcaf97dc