Remove short call timeout from func test
The bugfix I759aa36dc00a6c0612b9755dacd9aa414c408498 ensured that nova
rejectes a volume detach if the compute is down. But the func test still
has short rpc timeout set which sometimes causes that the server creates
times out on a slow node. Before the bugfix the short timeout was needed
to speed up the test but after the fix it is not needed any more so it
is removed to stabilize the test.

Change-Id: I12990eaca3820e56047e4d0e526c436fd2cfcf31
Related-Bug: #1909120