Add migration for 'internal_access_path' column to allow more than 255 chars.
The 'internal_access_path' column of 'console_auth_tokens' table has the
type String with max length of 255 characters which might not be enough
when used with VmWare and that will cause a DBDataError.

This migration changes the internal_access_path type to be Text with max
length of 65535 characters. And also adds a placeholder for migration
version number 391.

Change-Id: I4463f01ae727edd5e76b4a50860b116cbdea6124
Closes-Bug: #1900371