libvirt: Add parsing of firmware metadata files
Recent versions of libvirt provide the ability to auto-configure
firmware, using firmware metadata files provided by QEMU [1] to do so.
Unfortunately this feature is missing knobs to allow us to select e.g.
whether secure boot is enabled or what keys are enrolled [2]. Until
libvirt gains the ability to do this for us, we need to do it ourselves.

Add the ability to parse the firmware metadata files, providing a better
way to identify the location of firmware and NVRAM template files along
with the capabilities of same. This is pretty trivial thanks to the good
documentation that exists for same.

[1] https://github.com/qemu/qemu/blob/v5.2.0/docs/interop/firmware.json
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1906500

Change-Id: If08cbe2d5f2b301ac92a183f15fb87a18836bff5
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>