setup.cfg: Resolve warning
setuptools is now emitting the following warning when building nova:

  UserWarning: Usage of dash-separated 'python-requires' will not be
  supported in future versions. Please use the underscore name
  'python_requires' instead

Do as it says. While we're here, we update a few other options to use
their new variants and add some additional metadata. This leaves one
final warning:

  Warning: Unknown distribution option: 'requires_python'

That unfortunately cannot be fixed here and requires a separate fix in
pbr.

Change-Id: I859bc89bc1ee80652f471c3463ef6562d720263b
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>