[stable-only] gate: Pin CEPH_RELEASE to nautilus in LM hook
I1edd5a50079f325fa143a7e0d51b3aa3bb5ed45d moved the branchless
devstack-plugin-ceph project to the Octopus release of Ceph that drops
support for py2. As this was still the default on stable/train this
breaks the nova-live-migration and nova-grenade jobs.

This change works around this by pinning the CEPH_RELEASE to nautilus
within the LM hook as was previously used prior to the above landing.

Note that the devstack-plugin-ceph-tempest job from the plugin repo
continues to pass as it is correctly pinned to the Luminous release that
supports py2.

If anything the above enforces the need to move away from these hook
scripts and instead inherit our base ceph jobs from this repo in the
future to avoid the Ceph release jumping around like this.

Change-Id: I1d029ebe78b16ed2d4345201b515baf3701533d5
(cherry picked from commit ff570d1b4e9b9777405ae75cc09eae2ce255bf19)