[stable-only] Move grenade jobs to experimental
Grenade jobs in stable/pike are failing as stable/ocata's devstack
is broken and no fix is available. This patch moves out the grenade
jobs from check and gate jobs to experimental. In this way we are
not wasting resources, but still there is the option to run grenade
jobs in case ocata is somehow fixed.

Change-Id: I8fceeb6dd3ce1a1e888dfe68f4e0910009d506b3