doc: Use a non-numerical anchor when referencing microversions
As called out in the review for
Ib7b9f4fd7673525129c03dc2943deedd0c7ad81f the use of a numerical anchor
causes a sequence id to be used that can change over time in the
document and thus is not stable to reference externally.

This change simply switches to a non-numerical anchor ensuring that
sphinx generates a stable anchor we can always reference.

Change-Id: I550f7fd89a13e58031b0ddfbcb4f6a5239dbf335